Refactor the command line interface
Instead of having one CLI per tool, make the individual tools real
libraries and introduce a single new "cli" project. This has several
advantages:

- Tools now really are libraries from a build system point of view, and
  can be easily used in other projects like we were already advertising
  in the README.

- Using the new "java-library" Gradle plugin allows for significant build
  performance improvements by dividing dependencies into "api" and
  "implementation" dependencies, see [1].

- Command line parsing happens in only one place and code for global
  options is not duplicated anymore.

- The purpose of a tool is now also documented in the command line help
  as part of the (sub-)command description.

[1] https://docs.gradle.org/current/userguide/java_library_plugin.html#sec:java_library_recognizing_dependencies
[2] http://jcommander.org/#_more_complex_syntaxes_commands
