+++ analyzer/src/main/kotlin/managers/NPM.kt
@@ -24,9 +24,10 @@ import ch.frankel.slf4k.*
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.ObjectNode
 
-import com.here.ort.analyzer.Main
-import com.here.ort.analyzer.PackageManager
 import com.here.ort.analyzer.AbstractPackageManagerFactory
+import com.here.ort.analyzer.HTTP_CACHE_PATH
+import com.here.ort.analyzer.PackageManager
+import com.here.ort.analyzer.TOOL_NAME
 import com.here.ort.downloader.VersionControlSystem
 import com.here.ort.model.Error
 import com.here.ort.model.HashAlgorithm
@@ -140,7 +141,7 @@ open class NPM(analyzerConfig: AnalyzerConfiguration, repoConfig: RepositoryConf
             // Temporarily move away any existing "node_modules" directory within the same filesystem to ensure
             // the move can be performed atomically.
             if (modulesDir.isDirectory) {
-                val tempDir = createTempDir(Main.TOOL_NAME, ".tmp", workingDir)
+                val tempDir = createTempDir(TOOL_NAME, ".tmp", workingDir)
                 tempModulesDir = File(tempDir, "node_modules")
                 log.warn { "'$modulesDir' already exists, temporarily moving it to '$tempModulesDir'." }
                 Files.move(modulesDir.toPath(), tempModulesDir.toPath(), StandardCopyOption.ATOMIC_MOVE)
@@ -238,7 +239,7 @@ open class NPM(analyzerConfig: AnalyzerConfiguration, repoConfig: RepositoryConf
                     .url("https://registry.npmjs.org/$encodedName")
                     .build()
 
-            OkHttpClientHelper.execute(Main.HTTP_CACHE_PATH, pkgRequest).use { response ->
+            OkHttpClientHelper.execute(HTTP_CACHE_PATH, pkgRequest).use { response ->
                 if (response.code() == HttpURLConnection.HTTP_OK) {
                     log.debug {
                         if (response.cacheResponse() != null) {
