+++ downloader/src/funTest/kotlin/BabelTest.kt
@@ -76,7 +76,7 @@ class BabelTest : StringSpec() {
                     vcsProcessed = vcsMerged
             )
 
-            val downloadResult = Main.download(pkg, outputDir)
+            val downloadResult = Downloader().download(pkg, outputDir)
 
             downloadResult.sourceArtifact shouldBe null
             downloadResult.vcsInfo shouldNotBe null
