+++ downloader/src/funTest/kotlin/DirectoryTest.kt
@@ -64,7 +64,7 @@ class DirectoryTest : StringSpec() {
 
             // No download source specified, we expect exception in this case.
             shouldThrow<DownloadException> {
-                Main.download(pkg, outputDir)
+                Downloader().download(pkg, outputDir)
             }
 
             outputDir.list().size shouldBe 1
