+++ scanner/src/main/kotlin/Scanner.kt
@@ -19,7 +19,7 @@
 
 package com.here.ort.scanner
 
-import com.here.ort.downloader.consolidateProjectPackagesByVcs
+import com.here.ort.downloader.Downloader
 import com.here.ort.model.Environment
 import com.here.ort.model.OrtResult
 import com.here.ort.model.Package
@@ -34,6 +34,13 @@ import com.here.ort.model.readValue
 import java.io.File
 import java.util.ServiceLoader
 
+const val TOOL_NAME = "scanner"
+const val HTTP_CACHE_PATH = "$TOOL_NAME/cache/http"
+
+/**
+ * The class to run license / copyright scanners. The signatures of public functions in this class define the library
+ * API.
+ */
 abstract class Scanner(protected val config: ScannerConfiguration) {
     companion object {
         private val LOADER = ServiceLoader.load(ScannerFactory::class.java)!!
@@ -79,7 +86,7 @@ abstract class Scanner(protected val config: ScannerConfiguration) {
         val analyzerResult = ortResult.analyzer!!.result
 
         // Add the projects as packages to scan.
-        val consolidatedProjectPackageMap = consolidateProjectPackagesByVcs(analyzerResult.projects)
+        val consolidatedProjectPackageMap = Downloader().consolidateProjectPackagesByVcs(analyzerResult.projects)
         val consolidatedReferencePackages = consolidatedProjectPackageMap.keys.map { it.toCuratedPackage() }
 
         val projectScanScopes = if (scopesToScan.isNotEmpty()) {
