+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'oss-review-toolkit'
 
 include 'analyzer'
+include 'cli'
 include 'downloader'
 include 'model'
 include 'reporter'
