+++ utils/build.gradle
@@ -1,19 +1,22 @@
+// Apply core plugins.
+apply plugin: 'java-library'
+
 dependencies {
-    compile('ch.frankel.log4k:slf4k-api:1.0.0') {
+    api('ch.frankel.log4k:slf4k-api:1.0.0') {
         // Our version of the Kotlin runtime is provided by the Gradle plugin.
         exclude module: 'kotlin-runtime'
     }
 
     // Use logback-classic as the logger for slf4k as it allows to change the log level at runtime.
-    compile 'ch.qos.logback:logback-classic:1.2.3'
+    api 'ch.qos.logback:logback-classic:1.2.3'
 
-    compile "com.fasterxml.jackson.core:jackson-databind:$jacksonVersion"
+    api "com.fasterxml.jackson.core:jackson-databind:$jacksonVersion"
+    api "com.squareup.okhttp3:okhttp:$okhttpVersion"
+    api "com.vdurmont:semver4j:$semverVersion"
 
-    compile "com.jakewharton:disklrucache:$disklrucacheVersion"
-    compile "com.squareup.okhttp3:okhttp:$okhttpVersion"
-    compile "com.vdurmont:semver4j:$semverVersion"
+    implementation "com.jakewharton:disklrucache:$disklrucacheVersion"
 
-    compile "commons-codec:commons-codec:$commonsCodecVersion"
-    compile "org.apache.commons:commons-compress:$commonsCompressVersion"
-    compile "org.tukaani:xz:$xzVersion"
+    implementation "commons-codec:commons-codec:$commonsCodecVersion"
+    implementation "org.apache.commons:commons-compress:$commonsCompressVersion"
+    implementation "org.tukaani:xz:$xzVersion"
 }
